# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}/profile": {
      "patch": {
        "description": "Edita o perfil do próprio usuário. Apenas o próprio usuário pode editar seu perfil.",
        "operationId": "EditUserProfileController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário cujo perfil será editado",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dados do perfil a serem atualizados",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "João Silva",
                    "description": "Nome completo do usuário"
                  },
                  "position": {
                    "type": "string",
                    "example": "Técnico em Informática",
                    "description": "Cargo do usuário"
                  }
                },
                "required": ["name", "position"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão para editar este perfil"
          },
          "404": { "description": "Usuário não encontrado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Editar perfil do usuário",
        "tags": ["User Profile"]
      }
    },
...
    "/problems": {
      "post": {
        "description": "Cria um novo problema de infraestrutura",
        "operationId": "CreateProblemController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProblemRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Problema criado com sucesso" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Token JWT inválido ou expirado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Criar problema",
        "tags": ["Problems"]
      },
      "get": {
        "description": "Retorna uma lista paginada de problemas reportados no sistema com informações de localização",
        "operationId": "FetchProblemsController_handle",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "example": 1, "type": "number" }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Termo de busca para filtrar problemas por título ou descrição",
            "schema": { "example": "ar condicionado", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de problemas retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "problems": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProblemWithDetailsResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Parâmetros inválidos" }
        },
        "summary": "Buscar problemas",
        "tags": ["Problems"]
      }
    },
    "/problems/{slug}": {
      "get": {
        "description": "Retorna um problema específico pelo seu slug",
        "operationId": "GetProblemBySlugController_handle",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Slug do problema",
            "schema": {
              "example": "ar-condicionado-nao-funcionando",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Problema encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "problem": {
                      "$ref": "#/components/schemas/ProblemResponse"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Slug inválido" },
          "404": { "description": "Problema não encontrado" }
        },
        "summary": "Buscar problema por slug",
        "tags": ["Problems"]
      }
    },
    "/categories": {
      "get": {
        "description": "Retorna uma lista paginada de categorias de problemas disponíveis no sistema",
        "operationId": "FetchCategoriesController_handle",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "example": 1, "type": "number" }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluir categorias deletadas (na lixeira)",
            "schema": { "example": false, "type": "boolean" }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": { "example": true, "type": "boolean" }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Busca por nome ou descrição",
            "schema": { "example": "Climatização", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorias retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CategoryResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Parâmetros inválidos" }
        },
        "summary": "Buscar categorias",
        "tags": ["Categories"]
      },
      "post": {
        "description": "Cria uma nova categoria de problema (requer role MANAGER+)",
        "operationId": "CreateCategoryController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategoryRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Categoria criada com sucesso" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Criar categoria",
        "tags": ["Categories"]
      }
    },
...
    "/problems/{id}": {
      "put": {
        "description": "Edita um problema existente. Apenas o criador do problema pode editá-lo.",
        "operationId": "EditProblemController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do problema a ser editado",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditProblemRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Problema editado com sucesso" },
          "400": {
            "description": "Dados inválidos ou usuário não tem permissão"
          },
          "401": { "description": "Token JWT inválido ou expirado" },
          "404": { "description": "Problema não encontrado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Editar problema",
        "tags": ["Problems"]
      },
      "delete": {
        "description": "Deleta um problema do sistema. Apenas o criador do problema pode deletá-lo.",
        "operationId": "DeleteProblemController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do problema a ser deletado",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": { "description": "Problema deletado com sucesso" },
          "400": {
            "description": "Usuário não tem permissão ou problema não encontrado"
          },
          "401": { "description": "Token JWT inválido ou expirado" },
          "404": { "description": "Problema não encontrado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Deletar problema",
        "tags": ["Problems"]
      }
    }
  },
  "info": {
    "title": "IFAL Arapiraca Backend API",
    "description": "API para gerenciamento de problemas de infraestrutura do IFAL Arapiraca",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateProblemRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título do problema",
            "example": "Ar condicionado não está funcionando"
          },
          "description": {
            "type": "string",
            "description": "Descrição detalhada do problema",
            "example": "O ar condicionado da sala 201 não está ligando há 3 dias"
          },
          "categoryId": {
            "type": "string",
            "description": "ID da categoria do problema",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "locationId": {
            "type": "string",
            "description": "ID da localização onde ocorreu o problema",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        },
        "required": ["title", "description", "categoryId", "locationId"]
      },
      "CreateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da categoria",
            "example": "Climatização"
          },
          "description": {
            "type": "string",
            "description": "Descrição da categoria",
            "example": "Problemas relacionados a ar condicionado, ventilação e climatização"
          }
        },
        "required": ["name", "description"]
      },
...
      "EditProblemRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Novo título do problema",
            "example": "Ar condicionado com defeito na sala 201"
          },
          "description": {
            "type": "string",
            "description": "Nova descrição do problema",
            "example": "O ar condicionado da sala 201 está fazendo ruído estranho e não resfriando adequadamente"
          }
        },
        "required": ["title", "description"]
      }
    }
  }
}