# src/infra/http/controllers/fetch-categories.controller.ts
import { BadRequestException, Controller, Get, Query } from '@nestjs/common'
import { ApiTags, ApiOperation, ApiResponse, ApiQuery } from '@nestjs/swagger'
import { ZodValidationPipe } from '@/infra/http/pipes/zod-validation-pipe'
import { z } from 'zod'
import { FetchCategoriesUseCase } from '@/domain/maintenance-problems/application/use-cases/fetch-categories'
import { CategoryPresenter } from '../presenters/category-presenter'
import { CategoryResponse } from '../dtos/interfaces.dto'

const pageQueryParamSchema = z
  .string()
  .optional()
  .default('1')
  .transform(Number)
  .pipe(z.number().min(1))

const queryValidationPipe = new ZodValidationPipe(pageQueryParamSchema)

type PageQueryParamSchema = z.infer<typeof pageQueryParamSchema>

const querySearchParamSchema = z.string().optional()
const isActiveQueryParamSchema = z
  .string()
  .optional()
  .transform((val) => {
    if (val === undefined) return undefined
    return val === 'true'
  })
const includeDeletedQueryParamSchema = z
  .string()
  .optional()
  .transform((val) => {
    if (val === undefined) return undefined
    return val === 'true'
  })

const querySearchValidationPipe = new ZodValidationPipe(querySearchParamSchema)
const isActiveValidationPipe = new ZodValidationPipe(isActiveQueryParamSchema)
const includeDeletedValidationPipe = new ZodValidationPipe(
  includeDeletedQueryParamSchema,
)

type QuerySearchParamSchema = z.infer<typeof querySearchParamSchema>
type IsActiveQueryParamSchema = z.infer<typeof isActiveQueryParamSchema>
type IncludeDeletedQueryParamSchema = z.infer<
  typeof includeDeletedQueryParamSchema
>

@Controller('/categories')
@ApiTags('Categories')
export class FetchCategoriesController {
  constructor(private fetchCategories: FetchCategoriesUseCase) {}

  @Get()
  @ApiOperation({
    summary: 'Buscar categorias',
    description:
      'Retorna uma lista paginada de categorias de problemas disponíveis no sistema',
  })
  @ApiQuery({
    name: 'page',
    required: false,
    description: 'Número da página (começa em 1)',
    example: 1,
    type: Number,
  })
  @ApiQuery({
    name: 'query',
    required: false,
    description: 'Busca por nome ou descrição',
    example: 'Climatização',
    type: String,
  })
  @ApiQuery({
    name: 'isActive',
    required: false,
    description: 'Filtrar por status ativo/inativo',
    example: true,
    type: Boolean,
  })
  @ApiQuery({
    name: 'includeDeleted',
    required: false,
    description: 'Incluir categorias deletadas (na lixeira)',
    example: false,
    type: Boolean,
  })
  @ApiResponse({
    status: 200,
    description: 'Lista de categorias retornada com sucesso',
    schema: {
      type: 'object',
      properties: {
        categories: {
          type: 'array',
          items: { $ref: '#/components/schemas/CategoryResponse' },
        },
      },
    },
  })
  @ApiResponse({ status: 400, description: 'Parâmetros inválidos' })
  async handle(
    @Query('page', queryValidationPipe) page: PageQueryParamSchema,
    @Query('query', querySearchValidationPipe) query: QuerySearchParamSchema,
    @Query('isActive', isActiveValidationPipe)
    isActive: IsActiveQueryParamSchema,
    @Query('includeDeleted', includeDeletedValidationPipe)
    includeDeleted: IncludeDeletedQueryParamSchema,
  ) {
    const result = await this.fetchCategories.execute({
      page,
      query,
      isActive,
      includeDeleted,
    })

    if (result.isLeft()) {
      throw new BadRequestException()
    }

    const categories = result.value.categories

    return { categories: categories.map(CategoryPresenter.toHTTP) }
  }
}